+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixed
 
+- disallow putting already parenthesized elements to `Import/ImportFrom.names` and `With/AsyncWith.items`.
 - gave `AsyncWith` put one the correct handler, was missing cases because was doing normal expression put
 - `put_src(action='offset')` exactly on `FormattedValue` or `Interpolation` debug expression will update the preceding debug string correctly
 - fix `replace()` end of sequence with empty slice while using `one=False` returns `None` instead of raising